move all functionality into library target
