--- Sources/SwiftOutdated/PackageCollection.swift
@@ -1,7 +1,0 @@
-import Foundation
-import Outdated
-
-struct PackageCollection: Encodable {
-    var outdatedPackages: [OutdatedPackage]
-    var ignoredPackages: [SwiftPackage]
-}
